+++ src/solver.cpp
@@ -1537,7 +1537,7 @@ bool Solver::is_decision (int lit) {
   return res;
 }
 
-void Solver::force_backtrack (size_t new_level) {
+void Solver::force_backtrack (int new_level) {
   TRACE ("force_backtrack", new_level);
   REQUIRE_VALID_OR_SOLVING_STATE ();
   REQUIRE (external->propagator, "can not force backtrack without a connected propagator");
